duplicate bridge feature

+++ components/navbar.js
@@ -3,7 +3,7 @@ import { usePathname, useRouter } from 'next/navigation';
 import { useEffect, useRef, useState } from 'react';
 import { logoutUserFromMsg91, switchOrg, switchUser } from '@/config';
 import { useCustomSelector } from '@/customSelector/customSelector';
-import { deleteBridgeAction, getAllBridgesAction } from '@/store/action/bridgeAction';
+import { deleteBridgeAction, duplicateBridgeAction, getAllBridgesAction } from '@/store/action/bridgeAction';
 import { setCurrentOrgIdAction } from '@/store/action/orgAction';
 import { Building2, ChevronDown, FileSliders, History, Home, KeyRound, LogOut, Mail, Rss, Settings2 } from 'lucide-react';
 import { useDispatch } from 'react-redux';
@@ -162,27 +162,35 @@ function Navbar() {
     };
   }, []);
 
-  const handleDeleteBridge = async () => {
+  const handleDeleteBridge = async (item) => {
     const bridgeId = path[5];
-    const orgId = path[2];
-    // Confirm delete action
-    const confirmDelete = window.confirm('Are you sure you want to delete this bridge?');
 
-    // If confirmed
-    if (confirmDelete) {
-      try {
-        // Dispatch delete bridge action and get all bridges
-        dispatch(deleteBridgeAction({ bridgeId, orgId: orgId }));
-        router.push(`/org/${path[2]}/bridges`)
-        toast.success('Bridge deleted successfully');
-        dispatch(getAllBridgesAction());
-      } catch (error) {
-        // Log error
-        console.error('Failed to delete bridge:', error);
-        // Show toast error
-        toast.error('Error deleting bridge');
+    if (item.trim().toLowerCase() === 'duplicate') {
+      dispatch(duplicateBridgeAction(bridgeId));
+      toast.success('Bridge duplicate successfully');
+      router.push(`/org/${path[2]}/bridges`)
+    } else {
+      const orgId = path[2];
+      // Confirm delete action
+      const confirmDelete = window.confirm('Are you sure you want to delete this bridge?');
+
+      // If confirmed
+      if (confirmDelete) {
+        try {
+          // Dispatch delete bridge action and get all bridges
+          dispatch(deleteBridgeAction({ bridgeId, orgId: orgId }));
+          router.push(`/org/${path[2]}/bridges`)
+          toast.success('Bridge deleted successfully');
+          dispatch(getAllBridgesAction());
+        } catch (error) {
+          // Log error
+          console.error('Failed to delete bridge:', error);
+          // Show toast error
+          toast.error('Error deleting bridge');
+        }
       }
     }
+
   };
 
   return (
@@ -229,8 +237,8 @@ function Navbar() {
                 </svg>
               </div>
               <ul tabIndex={0} className="dropdown-content z-[9999999999] menu p-2 shadow bg-base-100 rounded-box w-52 custom-dropdown">
-                {['Delete'].map((item) => (
-                  <li key={item} onClick={handleDeleteBridge}>
+                {['Duplicate', 'Delete'].map((item) => (
+                  <li key={item} onClick={() => handleDeleteBridge(item)}>
                     <a className={path[3] === item ? "active" : ""}>{item.charAt(0).toUpperCase() + item.slice(1)}</a>
                   </li>
                 ))}
